Match Node Env in Heroku to Served Assets
For heroku, we have defined our node env as development. This feels a bit odd since we are serving production assets, so proposing we change the node env to be production as well. 
 
The drawback would be that by defining this env as production we will no longer see console warnings/messages in our deployments.